Combined results to a single folder
Things I changed:
- I moved this script into its own folder within utilities
- Results from running this script (screenshots + CSV) will be created in a timestamped file within this new folder
- Screenshots are numbered matching the corresponding index in the CSV file

I think what you had before for screenshots was overwriting itself as it was just returning the type of the data record, so there would only ever be one screenshot there. I figured that if we just number them to match the CSV, we can easily lookup which click belongs to which screenshot.