Implemented universal Window class + subclasses
Introduced a new class called Window that lets you control the game client properties, as well as access common points of interest across all OSRS-based games (with or without RuneLite).

This change adds a bit of overhead and boilerplate, but the benefits probably outweigh that. Additionally, we can now re-focus the window upon resuming a bot. With this new change, we can also easily implement template matching to find points of interest if it's necessary.