Spellbook checker
Added method to main bot.py that returns what spellbook we're currently on.

Does this by trying to search for the spellbook sprites on the main client screen.
The spellbook image paths are stored in an enum so that we can more cleanly compare which book is returned.

Also renamed the src/images/bot/spellbooks/normal directory to standard as this is how this spellbook is referenced in game and by the wiki.