Adding Pushbullet integration
This allows a call to the send_notification function of bot.py
so that updates can be sent to whatever device you have recieving
pushbullet notifications. For example, a good use case would be for bot
developers add a notification hook when we have run out of runes
or resources when high alching.

The API key is configured via settings and is unique to the bot user.
The pushbullet API key (Access token) can be found at
https://www.pushbullet.com/#settings/account.

If the API key is not configured, bot scripts will continue as normal.