Update chatbox_text located in bot.py
I added some functionality to add colors into the arguments to check if something is(TRUE) or isn't(FALSE) in the chatbox. Originally it only had functionality for player chat and based on the description it was destined for more but obviously a low priority. Now you can add a color argument to check in game messages as true or false.